Moved Status message helper functions closer to their uses
