+++ backend/src/ws/ws_structs/mod.rs
@@ -14,24 +14,3 @@ pub trait WsMessageWrapper: Serialize {
         serde_json::to_string(self).unwrap()
     }
 }
-
-fn get_finished_player_names(game: &Game) -> Vec<String> {
-    game.get_finished_players()
-        .iter()
-        .map(|p| p.name())
-        .collect()
-}
-
-fn find_author_name(game: &Game) -> Result<String, CreateStatusError> {
-    match game.find_author() {
-        None => Err(CreateStatusError::AuthorNotFound),
-        Some(author) => Ok(author.name()),
-    }
-}
-
-fn get_current_player_name(game: &Game) -> Result<String, CreateStatusError> {
-    match game.get_current_player() {
-        None => Err(CreateStatusError::CurrentPlayerNotFound),
-        Some(player) => Ok(player.name()),
-    }
-}
